UnifiedMap: Fix crash on resuming Google Maps (fix #16180) Move "repaintRotationIndicator" from UnifiedMapActivity to AbstractMapFragment
